Add files via upload
The effectiveness of cancer prediction system helps the people to know their cancer risk with low cost and it also helps the people to take the appropriate decision based on their cancer risk status. This dataset contains information on patients with lung cancer, including their age, Smokes or not, Area ,Alcoholic or not .Based on these details we Predict the status of the patient by using K Neighbors classification algorithm.